rm state conversion
