+++ keccak256/src/permutation.rs
@@ -10,7 +10,6 @@ pub(crate) mod pi;
 pub(crate) mod rho;
 pub(crate) mod rho_checks;
 pub(crate) mod rho_helpers;
-pub(crate) mod state_conversion;
 pub(crate) mod tables;
 pub(crate) mod theta;
 pub(crate) mod xi;
