--- keccak256/src/permutation/state_conversion.rs
@@ -1,209 +1,0 @@
-use halo2_proofs::{
-    circuit::{AssignedCell, Layouter},
-    plonk::{Advice, Column, ConstraintSystem, Error},
-};
-
-use crate::permutation::base_conversion::BaseConversionConfig;
-use crate::permutation::tables::BaseInfo;
-use eth_types::Field;
-use std::convert::TryInto;
-
-#[derive(Debug, Clone)]
-pub(crate) struct StateBaseConversion<F> {
-    bcc: BaseConversionConfig<F>,
-}
-
-impl<F: Field> StateBaseConversion<F> {
-    /// Side effect: parent flag is enabled
-    pub(crate) fn configure(
-        meta: &mut ConstraintSystem<F>,
-        bi: BaseInfo<F>,
-        lane: Column<Advice>,
-        flag: Column<Advice>,
-    ) -> Self {
-        meta.enable_equality(flag);
-        let bcc = BaseConversionConfig::configure(meta, bi, lane, flag);
-
-        Self { bcc }
-    }
-
-    pub(crate) fn assign_region(
-        &self,
-        layouter: &mut impl Layouter<F>,
-        state: &[AssignedCell<F, F>; 25],
-        flag: AssignedCell<F, F>,
-    ) -> Result<[AssignedCell<F, F>; 25], Error> {
-        let state: Result<Vec<AssignedCell<F, F>>, Error> = state
-            .iter()
-            .map(|lane| {
-                let output = self
-                    .bcc
-                    .assign_region(layouter, lane.clone(), flag.clone())?;
-                Ok(output)
-            })
-            .into_iter()
-            .collect();
-        let state = state?;
-        let state: [AssignedCell<F, F>; 25] = state.try_into().unwrap();
-        Ok(state)
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::arith_helpers::convert_b2_to_b13;
-    use crate::gate_helpers::biguint_to_f;
-    use crate::permutation::tables::FromBinaryTableConfig;
-    use halo2_proofs::pairing::bn256::Fr as Fp;
-    use halo2_proofs::{
-        circuit::{Layouter, SimpleFloorPlanner},
-        dev::MockProver,
-        plonk::{Advice, Circuit, Column, ConstraintSystem, Error},
-    };
-    use pretty_assertions::assert_eq;
-    #[test]
-    fn test_state_base_conversion() {
-        // We have to use a MyConfig because:
-        // We need to load the table
-        #[derive(Debug, Clone)]
-        struct MyConfig<F> {
-            flag: Column<Advice>,
-            state: [Column<Advice>; 25],
-            table: FromBinaryTableConfig<F>,
-            conversion: StateBaseConversion<F>,
-        }
-        impl<F: Field> MyConfig<F> {
-            pub fn configure(meta: &mut ConstraintSystem<F>) -> Self {
-                let table = FromBinaryTableConfig::configure(meta);
-                let state: [Column<Advice>; 25] = (0..25)
-                    .map(|_| meta.advice_column())
-                    .collect::<Vec<_>>()
-                    .try_into()
-                    .unwrap();
-                let flag = meta.advice_column();
-                let lane = meta.advice_column();
-                let bi = table.get_base_info(false);
-                let conversion = StateBaseConversion::configure(meta, bi, lane, flag);
-                Self {
-                    flag,
-                    state,
-                    table,
-                    conversion,
-                }
-            }
-
-            pub fn load(&self, layouter: &mut impl Layouter<F>) -> Result<(), Error> {
-                self.table.load(layouter)
-            }
-
-            pub fn assign_region(
-                &self,
-                layouter: &mut impl Layouter<F>,
-                input: [F; 25],
-            ) -> Result<[F; 25], Error> {
-                let flag_value = F::one();
-                let (state, flag) = layouter.assign_region(
-                    || "Input state",
-                    |mut region| {
-                        let state: [AssignedCell<F, F>; 25] = input
-                            .iter()
-                            .enumerate()
-                            .map(|(idx, &value)| {
-                                region
-                                    .assign_advice(
-                                        || format!("State {}", idx),
-                                        self.state[idx],
-                                        0,
-                                        || Ok(value),
-                                    )
-                                    .unwrap()
-                            })
-                            .collect::<Vec<_>>()
-                            .try_into()
-                            .unwrap();
-                        let flag =
-                            region.assign_advice(|| "Flag", self.flag, 0, || Ok(flag_value))?;
-                        Ok((state, flag))
-                    },
-                )?;
-                let output_state = self.conversion.assign_region(layouter, &state, flag)?;
-                let output_state: [F; 25] = output_state
-                    .iter()
-                    .map(|cell| cell.value().copied().unwrap_or_default())
-                    .collect::<Vec<F>>()
-                    .try_into()
-                    .unwrap();
-                Ok(output_state)
-            }
-        }
-
-        #[derive(Default)]
-        struct MyCircuit<F> {
-            in_state: [F; 25],
-            out_state: [F; 25],
-        }
-        impl<F: Field> Circuit<F> for MyCircuit<F> {
-            type Config = MyConfig<F>;
-            type FloorPlanner = SimpleFloorPlanner;
-
-            fn without_witnesses(&self) -> Self {
-                Self::default()
-            }
-
-            fn configure(meta: &mut ConstraintSystem<F>) -> Self::Config {
-                Self::Config::configure(meta)
-            }
-
-            fn synthesize(
-                &self,
-                config: Self::Config,
-                mut layouter: impl Layouter<F>,
-            ) -> Result<(), Error> {
-                config.load(&mut layouter)?;
-                let out_state = config.assign_region(&mut layouter, self.in_state)?;
-                assert_eq!(out_state, self.out_state);
-                Ok(())
-            }
-        }
-        let in_state: [[u64; 5]; 5] = [
-            [4398046511105, 8, 2, 268436480, 2305844108725321728],
-            [
-                17592186044416,
-                52776560230400,
-                544,
-                68719493120,
-                2199023255552,
-            ],
-            [
-                4398046543872,
-                1152921504606846984,
-                262144,
-                1024,
-                1099511627780,
-            ],
-            [0, 52776558133248, 514, 268451840, 2305845208236949504],
-            [17592186077184, 1152921504608944128, 262176, 68719476736, 4],
-        ];
-
-        let in_state_flat = in_state.iter().flatten().collect::<Vec<_>>();
-        let in_state: [Fp; 25] = in_state_flat
-            .iter()
-            .map(|&x| Fp::from(*x))
-            .collect::<Vec<_>>()
-            .try_into()
-            .unwrap();
-        let out_state: [Fp; 25] = in_state_flat
-            .iter()
-            .map(|&x| biguint_to_f::<Fp>(&convert_b2_to_b13(*x)))
-            .collect::<Vec<_>>()
-            .try_into()
-            .unwrap();
-        let circuit = MyCircuit::<Fp> {
-            in_state,
-            out_state,
-        };
-        let prover = MockProver::<Fp>::run(17, &circuit, vec![]).unwrap();
-        assert_eq!(prover.verify(), Ok(()));
-    }
-}
